+++ internal/product/model.go
@@ -7,10 +7,10 @@ type Product struct {
 	Name      string    `json:"name"`
 	Amount    int64     `json:"amount"`
 	Unit      string    `json:"unit"`
-	Calories  int64     `json:"calories"`
-	Protein   int64     `json:"protein"`
-	Fat       int64     `json:"fat"`
-	Carbs     int64     `json:"carbs"`
+	Calories  float64   `json:"calories"`
+	Protein   float64   `json:"protein"`
+	Fat       float64   `json:"fat"`
+	Carbs     float64   `json:"carbs"`
 	UserId    string    `json:"-"`
 	FitId     string    `json:"-"`
 	CreatedAt time.Time `json:"-"`
@@ -18,13 +18,13 @@ type Product struct {
 }
 
 type ProductCreate struct {
-	Name     string `json:"name"`
-	Amount   int64  `json:"amount"`
-	Unit     string `json:"unit"`
-	Calories int64  `json:"calories"`
-	Protein  int64  `json:"protein"`
-	Fat      int64  `json:"fat"`
-	Carbs    int64  `json:"carbs"`
-	UserId   string `json:"-"`
-	FitId    string `json:"-"`
+	Name     string  `json:"name"`
+	Amount   int64   `json:"amount"`
+	Unit     string  `json:"unit"`
+	Calories float64 `json:"calories"`
+	Protein  float64 `json:"protein"`
+	Fat      float64 `json:"fat"`
+	Carbs    float64 `json:"carbs"`
+	UserId   string  `json:"-"`
+	FitId    string  `json:"-"`
 }
